Fix for invalid file upload
see: https://gerrit.wikimedia.org/r/#/c/192303/2

If a user selects a file that is not allowed by $wgFileExtensions and
tries to upload it (without noticing the message)
 $uploadForm->mDesiredDestName
is an empty string. In this case NSFileRepo causes a fatal error.
Jenkins breaks on this commit [1] because of the extensions dependency to "Lockdown". Therefor I'll do a manual verification.
[1] https://integration.wikimedia.org/ci/job/mwext-NSFileRepo-testextension-zend/3/console